Core/Mail: fix cash on delivery charges above 214,748 gold
The cash on delivery payment was cast to a signed 32-bit integer before being taken from the recipient. Above 214,748 gold the value overflowed: for 250,000 gold the recipient gained about 179,000 gold instead of paying, while the sender did receive the 250,000 gold. Two players could create gold this way. The amount is now taken on 64 bits.

+++ src/server/game/Handlers/MailHandler.cpp
@@ -482,7 +482,7 @@ void WorldSession::HandleMailTakeItem(WorldPackets::Mail::MailTakeItem& packet)
             if (receive || sender_accId)
                 MailDraft(m->subject, "") .AddMoney(m->COD) .SendMailTo(trans, MailReceiver(receive, m->sender), MailSender(MAIL_NORMAL, m->receiver), MAIL_CHECK_MASK_COD_PAYMENT);
 
-            player->ModifyMoney(-int32(m->COD));
+            player->ModifyMoney(-int64(m->COD));
         }
 
         m->COD = 0;
